+++ myPacManProject/src/pacman/entries/pacman/neuralPacMan/nodes/Node.java
@@ -3,5 +3,5 @@ package pacman.entries.pacman.neuralPacMan.nodes;
 import pacman.game.Game;
 
 public interface Node {
-	public float value(int pacManIndex, Game game);
+	public double value(int pacManIndex, Game game);
 }
